array: Organizing utility functions.
* Added `absolute_slice` slicing.
